\name{autovarCore-package}
\alias{autovarCore-package}
\docType{package}
\title{
Automated Vector Autoregression Networks
}
\description{
Find contemporaneous and dynamic relations in time series data and return them as a graph.
}
\details{
\tabular{ll}{
Package: \tab autovarCore\cr
Type: \tab Package\cr
Version: \tab 1.0-0\cr
Date: \tab 2015-03-03\cr
License: \tab Unlimited\cr
}
Please see help of the \code{\link{autovar}} function for information on how to use this package.
}
\author{
Ando Emerencia

Maintainer: Ando Emerencia <ando.emerencia@gmail.com>
}
\keyword{ vector autoregression }
\keyword{ timeseries }
\keyword{ contemporaneous correlation }
\keyword{ granger causality }
