% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coefficients_of_kurtosis}
\alias{coefficients_of_kurtosis}
\title{Kurtosis coefficients.}
\usage{
coefficients_of_kurtosis(matrix)
}
\arguments{
\item{matrix}{the matrix of residuals.}
}
\description{
Kurtosis coefficients.
}

