% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coefficients_of_skewness}
\alias{coefficients_of_skewness}
\title{Skewness coefficients.}
\usage{
coefficients_of_skewness(matrix)
}
\arguments{
\item{matrix}{the matrix of residuals.}
}
\description{
Skewness coefficients.
}

