% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compete.r
\name{compete}
\alias{compete}
\title{Returns the winning model}
\usage{
compete(best, challenger, compare_outliers)
}
\arguments{
\item{best}{A model given as a list with at least the properties \code{varest, model_score,} and \code{bucket}.}

\item{challenger}{Another model, also given as a list with properties \code{varest, model_score,} and \code{bucket}.}

\item{compare_outliers}{A boolean. When \code{FALSE}, the model comparison does not take the number of dummy variables into account.}
}
\value{
This function returns the best model of the two given models.
}
\description{
This function returns the best model as explained in the documentation for the \code{autovar} function.
}
\examples{
data_matrix <- matrix(nrow = 40, ncol = 3)
data_matrix[, ] <- runif(ncol(data_matrix) * nrow(data_matrix), 1, nrow(data_matrix))
colnames(data_matrix) <- c('rumination', 'happiness', 'activity')
varest1 <- autovarCore:::run_var(data_matrix, NULL, 1)
model1 <- list(logtransformed = FALSE, lag = 1, varest = varest1, model_score = 100, bucket = 0.05)
varest2 <- autovarCore:::run_var(data_matrix, NULL, 2)
model2 <- list(logtransformed = FALSE, lag = 2, varest = varest2, model_score = 200, bucket = 0.01)
autovarCore:::compete(model1, model2, TRUE)
}

