% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/day_dummies.r
\name{day_dummies}
\alias{day_dummies}
\title{Calculate day dummy variables}
\usage{
day_dummies(number_of_rows, measurements_per_day)
}
\arguments{
\item{number_of_rows}{the number of rows in the input data set.}

\item{measurements_per_day}{the number of measurements per day in the input data set.}
}
\value{
Either \code{NULL} or a matrix with \code{number_of_rows} rows and \code{6} columns.
}
\description{
This function returns either \code{NULL} (if \code{measurements_per_day} is 0) or a matrix of daydummy variables specified number of rows and measurements per day. In the latter case, we return a matrix of six columns.
}
\examples{
autovarCore:::day_dummies(16, 2)
}

