% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/daypart_dummies.r
\name{daypart_dummies}
\alias{daypart_dummies}
\title{Calculate day-part dummy variables}
\usage{
daypart_dummies(number_of_rows, measurements_per_day)
}
\arguments{
\item{number_of_rows}{the number of rows in the input data set.}

\item{measurements_per_day}{the number of measurements per day in the input data set.}
}
\value{
Either \code{NULL} or a matrix with \code{number_of_rows} rows and \code{measurements_per_day - 1} columns.
}
\description{
This function returns either \code{NULL} (if \code{measurements_per_day} is 0 or 1) or a matrix of dummy variables for the specified input configuration.
}
\examples{
autovarCore:::daypart_dummies(10, 3)
}

