% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invalid_mask.r
\name{invalid_mask}
\alias{invalid_mask}
\title{Calculate a bit mask to identify invalid outlier dummies}
\usage{
invalid_mask(outlier_dummies)
}
\arguments{
\item{outlier_dummies}{A matrix of outlier dummy variables in columns.}
}
\value{
An integer mask indicating the invalid columns according to the procedure describe above.
}
\description{
Invalid outlier dummy variables are dummy variables that are all zeros (where the original variable had no outliers at the 2.5x standard deviation for either the residuals or the squared residuals. Interpreting the leftmost column as bit 0 and continuing with higher bits going from left to right in the matrix, this function returns a bit mask that has a 1 on all positions in the matrix where the dummy column is invalid. We use this in later functions to easily filter out theinvalid outlier masks from the valid ones.
}
\examples{
resid_matrix <- matrix(rnorm(39 * 3),
                       nrow = 39,
                       ncol = 3,
                       dimnames = list(NULL, c('rumination', 'happiness', 'activity')))
outlier_dummies <- autovarCore:::residual_outliers(resid_matrix, 40)
autovarCore:::invalid_mask(outlier_dummies)
}

