% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{portmanteau_test_statistics}
\alias{portmanteau_test_statistics}
\title{A faster version of portmanteau_test_statistic.}
\usage{
portmanteau_test_statistics(matrix)
}
\arguments{
\item{matrix}{the matrix of residuals or squared residuals.}
}
\description{
A faster version of portmanteau_test_statistic.
}

