% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print_correlation_matrix.r
\name{print_correlation_matrix}
\alias{print_correlation_matrix}
\title{Print the correlation matrix of the residuals of a model annotated with p-values}
\usage{
print_correlation_matrix(varest)
}
\arguments{
\item{varest}{A \code{varest} model.}
}
\value{
This function returns the annotated correlation matrix.
}
\description{
This function prints the correlation matrix of residuals of a model annotated with p-values.
}
\examples{
data_matrix <- matrix(nrow = 40, ncol = 3)
data_matrix[, ] <- runif(ncol(data_matrix) * nrow(data_matrix), 1, nrow(data_matrix))
colnames(data_matrix) <- c('rumination', 'happiness', 'activity')
varest <- autovarCore:::run_var(data_matrix, NULL, 1)
autovarCore::print_correlation_matrix(varest)
}

