% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selected_columns.r
\name{selected_columns}
\alias{selected_columns}
\title{Convert an outlier_mask to a vector of column indices}
\usage{
selected_columns(outlier_mask)
}
\arguments{
\item{outlier_mask}{An integer representing the outlier mask.}
}
\value{
A vector of column indices corresponding to the outlier mask.
}
\description{
This function returns an ordered vector of all the 1-toggled bits in the outlier_mask offset by 1.
}
\examples{
outlier_mask <- 7
autovarCore:::selected_columns(outlier_mask)
}

