% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_joint_sktest.r
\name{sktest_joint_p}
\alias{sktest_joint_p}
\title{SK test p-level}
\usage{
sktest_joint_p(Z1, Z2, n)
}
\arguments{
\item{Z1}{The Z score for skewness.}

\item{Z2}{The Z score for kurtosis.}

\item{n}{The number of rows in the residuals column.}
}
\description{
SK test p-level
}
