% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trend_columns.r
\name{trend_columns}
\alias{trend_columns}
\title{Construct linear and quadratic trend columns}
\usage{
trend_columns(number_of_rows)
}
\arguments{
\item{number_of_rows}{the number of rows in the input data set.}
}
\value{
A matrix with \code{number_of_rows} rows and 2 columns, one for linear trends and one for quadratic trends.
}
\description{
This function returns a matrix of linear and quadratic trends.
}
\examples{
autovarCore:::trend_columns(10)
}

