% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{portmanteau_test_statistics}
\alias{portmanteau_test_statistics}
\title{An implementation of the portmanteau test.}
\usage{
portmanteau_test_statistics(matrix)
}
\arguments{
\item{matrix}{the matrix of residuals or squared residuals.}
}
\description{
See the paper of Ljung-Box test for the used definition of autocorrelation.
}
